refactor: rename components and update imports for consistency
